backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
supabase
pydantic[email]